StudentManagementSystem/requirements.txt
amqp==5.3.1
asgiref==3.6.0
async-timeout==5.0.1
attrs==23.1.0
autobahn==23.6.1
Automat==22.10.0
billiard==4.2.1
buildozer==1.5.0
celery==5.4.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==15.1.0
cron-descriptor==1.4.5
cryptography==41.0.1
daphne==4.0.0
defusedxml==0.8.0rc2
distlib==0.3.6
Django==4.2.7
django-celery-beat==2.7.0
django-channels==0.7.0
django-cors-headers==3.14.0
django-countries==7.6.1
django-debug-toolbar==4.0.0
django-etc==1.4.0
django-filter==23.2
django-google-maps==0.13.0
django-hitcount==1.3.5
django-ranged-response==0.2.0
django-redis==5.4.0
django-simple-captcha==0.5.17
django-timezone-field==7.0
django-tracking2==0.5.1
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.11.1
filelock==3.10.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
mysql-connector-python==8.0.32
numpy==1.24.1
oauthlib==3.2.2
pexpect==4.8.0
Pillow==9.5.0
platformdirs==3.1.1
prompt_toolkit==3.0.48
protobuf==3.20.3
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pygame==2.3.0
PyJWT==2.8.0
pyOpenSSL==23.2.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
referencing==0.35.1
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.21.0
service-identity==21.1.0
sh==2.0.3
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.3
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.0.2
vine==5.1.0
virtualenv==20.21.0
wcwidth==0.2.13
zope.interface==6.0
